Remove configurable User-implementation
