--- controllers/Users/AllowedUser.php
@@ -1,14 +1,0 @@
-<?php
-
-
-namespace Grocy\Controllers\Users;
-
-
-class AllowedUser extends User
-{
-
-    public function hasPermission(string $permission): bool
-    {
-        return true;
-    }
-}
