--- controllers/Users/LockedUser.php
@@ -1,14 +1,0 @@
-<?php
-
-
-namespace Grocy\Controllers\Users;
-
-
-class LockedUser extends User
-{
-
-    public function hasPermission(string $permission): bool
-    {
-        return false;
-    }
-}
